assisted-vision-goggles/assisted_vision_goggle/src/requirements.txt
opencv-python
numpy==2.0.2
easyocr==1.7.2
ultralytics==8.3.22
requests==2.32.3
Pillow==10.4.0
mediapipe==0.10.14
pyttsx3==2.90
